+++ postgres-types/CHANGELOG.md
@@ -1,8 +1,45 @@
 # Change Log
 
+## Unreleased
+
+## v0.2.8 - 2024-09-15
+
+### Added
+
+* Added support for `jiff` 0.1 via the `with-jiff-01` feature.
+
+## v0.2.7 - 2024-07-21
+
+### Added
+
+* Added `Default` implementation for `Json`.
+* Added a `js` feature for WASM compatibility.
+
+### Changed
+
+* `FromStr` implementation for `PgLsn` no longer allocates a `Vec` when splitting an lsn string on it's `/`.
+* The `eui48-1` feature no longer enables default features of the `eui48` library.
+
+## v0.2.6 - 2023-08-19
+
+### Fixed
+
+* Fixed serialization to `OIDVECTOR` and `INT2VECTOR`.
+
+### Added
+
+* Removed the `'static` requirement for the `impl BorrowToSql for Box<dyn ToSql>`.
+* Added a `ToSql` implementation for `Cow<[u8]>`.
+
+## v0.2.5 - 2023-03-27
+
+### Added
+
+* Added support for multi-range types.
+
 ## v0.2.4 - 2022-08-20
 
-## Added
+### Added
 
 * Added `ToSql` and `FromSql` implementations for `Box<[T]>`.
 * Added `ToSql` and `FromSql` implementations for `[u8; N]` via the `array-impls` feature.